finish property panel ui part mode
